Fixes AppSync API resource attribute references.
Corrects incorrect resource attribute references within the
CloudFormation template for the AppSync API. Specifically, it
resolves issues where the `Arn` and `ApiId` attributes were not
correctly referenced, leading to deployment failures. This change
ensures that resources are properly associated with the AppSync API
during stack creation and updates.

Also adds a missing condition for the CloudFront response headers policy
so it deploys only if the web ui is deployed.